(Breaking change): Refactor from ArtsyAuth::ApplicationController base class to ArtsyAuth::Authenticated module
