+++ lib/artsy-auth.rb
@@ -1,8 +1,8 @@
+require 'artsy-auth/authenticated'
 require 'artsy-auth/config'
 require 'artsy-auth/engine'
-require 'artsy-auth/version'
-require 'artsy-auth/application_controller'
 require 'artsy-auth/session_controller'
+require 'artsy-auth/version'
 
 module ArtsyAuth
 end
