+++ spec/dummy/app/controllers/unimplemented_controller.rb
@@ -1,0 +1,7 @@
+class UnimplementedController < ApplicationController
+  include ArtsyAuth::Authenticated
+
+  def index
+    render plain: 'Hello from unimplemented world'
+  end
+end
