+++ spec/dummy/config/routes.rb
@@ -1,3 +1,6 @@
 Rails.application.routes.draw do
   mount ArtsyAuth::Engine => '/'
+  get 'private' => 'private#index'
+  get 'public' => 'application#index'
+  get 'unimplemented' => 'unimplemented#index'
 end
